% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatBismark.R
\name{concatBismark}
\alias{concatBismark}
\title{Merge Bismark coverage files
This function merge a list of Bismark coverage files by adding methylation values}
\usage{
concatBismark(listbismark = list)
}
\arguments{
\item{list}{list of bismark files}
}
\value{
A dataframe with merged genomic locations and added methylation values
}
\description{
Merge Bismark coverage files
This function merge a list of Bismark coverage files by adding methylation values
}
\examples{
concatBismark(list(bedA, bedB, bedC))
}
\keyword{bismark}
\keyword{merge}
\keyword{methylation}
\keyword{union}
