% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_btw_clusters.R
\name{nn_bw_clusters}
\alias{nn_bw_clusters}
\title{Next-neighbors in different clusters}
\usage{
nn_bw_clusters(seurat_object, top_nn = 10)
}
\arguments{
\item{seurat_object}{Dataframe with seurat marker predictions.}

\item{top_nn}{use mouse or human (default: "mouse")}
}
\value{
array with cells sharing neighbors in different clusters
}
\description{
Next-neighbors in different clusters
}
\examples{
nn_bw_clusters(pbmc, top_nn = 10)
}
\keyword{clusters}
\keyword{next-neighbors}
\keyword{single-cell}
