% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGeneList.R
\name{randomGeneList}
\alias{randomGeneList}
\title{Random gene list
This function creates random gene lists.}
\usage{
randomGeneList(n_genes = 10)
}
\arguments{
\item{n_genes}{How many random genes? Defaults to 10.}
}
\value{
A 2 column table with gene names in the second column (gene).
}
\description{
Random gene list
This function creates random gene lists.
}
\examples{
randomGeneList(n_genes=10)
}
\keyword{genes}
